+++ Changelog.md
@@ -24,6 +24,7 @@ and this project adheres to
 -   mAh Calculated at twice the rate
 -   POF Error when not profiling
 -   UI Improvements
+-   Update active battery profile after download
 
 ## 0.0.14 - 2023-04-28